I had to tidy up the GameManager class, I edited the .h file but forgot to update the .cpp as well. After that I made the regions of code more explicit and tweaked the game.cpp file very slightly. Next I'll start working on the Elf class

+++ cocos/Classes/GameManager.h
@@ -14,15 +14,21 @@ public:
 	GameManager();
 	~GameManager();
 
+	// Scoring
+	//======================================================================
 	int GetScore();
 	void AddToScore(int increment);
 	void ResetScore();
 
+	// Elves
+	//======================================================================
 	void SetNumOfElfs(int numOfElfs);
 	int GetNumOfElfs();
 
+	// Speed
+	//======================================================================
 	void SetSpeed(float newSpeed);
-	int GetSpeed();
+	float GetSpeed();
 };
 
 #endif // __GAMEMANAGER_H__
